+++ src/main/java/de/sakurajin/evenbetterarcheology/api/owo_annotations/ModdedRarity.java
@@ -1,0 +1,8 @@
+package de.sakurajin.evenbetterarcheology.api.owo_annotations;
+
+public enum ModdedRarity {
+    COMMON,
+    UNCOMMON,
+    RARE,
+    EPIC
+}