+++ src/main/java/de/sakurajin/evenbetterarcheology/item/ModItems.java
@@ -6,14 +6,15 @@ import io.wispforest.owo.itemgroup.OwoItemSettings;
 import io.wispforest.owo.registration.annotations.AssignedName;
 import io.wispforest.owo.registration.reflect.ItemRegistryContainer;
 import net.minecraft.item.Item;
+import net.minecraft.item.Items;
 import net.minecraft.util.Rarity;
 
 public class ModItems implements ItemRegistryContainer {
     //ITEM ENTRIES-------------------------------------------------------------------------//
     //BRUSHES
-    public static final Item IRON_BRUSH = BetterBrushItem.Builder().setBrushingSpeed(8).setMaxDamage(128).build();
+    public static final Item IRON_BRUSH = BetterBrushItem.Builder().setBrushingSpeed(8).setMaxDamage(128).setMaterial(Items.IRON_INGOT).build();
 
-    public static final Item DIAMOND_BRUSH = BetterBrushItem.Builder().setBrushingSpeed(6).setMaxDamage(256).build();
+    public static final Item DIAMOND_BRUSH = BetterBrushItem.Builder().setBrushingSpeed(6).setMaxDamage(256).setMaterial(Items.DIAMOND).build();
 
     public static final Item ARTIFACT_SHARDS = new Item(new OwoItemSettings().group(EvenBetterArcheology.GROUP).rarity(Rarity.UNCOMMON));